+++ src/main/java/ru/yandex/practicum/filmorate/model/FriendShip.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class FriendShip {
-    private long userId;
-    private long friendId;
+    private Integer userId;
+    private Integer friendId;
     private String status;
 }
